added some more tests

+++ include/settings/GridSettings.h
@@ -9,7 +9,7 @@ namespace settings {
         extern double scaling;       // The percent increase in grid size in all dimensions when the grid size is automatically deduced based on an input vector of atoms.
         extern bool cubic;           // Whether to generate a cubic grid. This is primarily intended for rigid body optimization, to ensure there's enough space for all possible conformations.
         extern double rvol;          // The radius of the excluded volume sphere around each atom.
-        extern int exv_radius;       // The radius of the excluded volume sphere used for the grid-based excluded volume calculations in Å.
+        extern double exv_radius;    // The radius of the excluded volume sphere used for the grid-based excluded volume calculations in Å.
         extern bool save_exv;        // Whether to save the excluded volume grid when using the grid-based excluded volume calculations.
 
         namespace detail {
